% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pp.prototypes.r
\docType{package}
\name{pp.prototypes}
\alias{pp.prototypes}
\alias{pp.prototypes-package}
\title{Prototype data for analysis of public pension plans.}
\description{
Currently includes:\cr
\code{\link{actives}} Dataframe of actives for prototypical public pension plans\cr
\code{\link{retirees}} Dataframe of retirees for prototypical public pension plans\cr
\code{\link{salgrowth.hist}} Dataframe with historical salary growth by age\cr
\code{\link{salgrowth.assume}} Dataframe with assumed salary growth by age\cr
}

